Create Program
In this homework, you will work with some different variable types, branches, and loops.
You have to write a program to draw figures using an ASCII character of the user's choice. The program starts with a welcome message:
Hello, welcome to Figures Drawing!
Then, the user is presented with a menu of options:
Select a figure from the menu (or quit):
1 - Square
2 - Rectangle
3 - Triangle
9 - Quit program
Option 1: if the user enters 1, they wish to draw a square. Ask them for a character to be used to draw the figure:
Enter a character:
Then, ask for a side:
Enter side length:
Use this information to draw the square. For example, if the user entered the character Y and the number 5, the square will look like this (the proportions may be off, but what matters is the number of characters):
YYYYY
YYYYY
YYYYY
YYYYY
YYYYY
Option 2: if the user enters 2, they wish to draw a rectangle. Ask them for a character to be used to draw the figure:
Enter a character:
Then, ask for a height and width:
Enter height:
Enter width:
Remember to save both input immediately after they are entered by the user. Use this information to draw the rectangle. For example, if the user entered the character * and the numbers 2 and 6, the rectangle will look like this:
******
******
Option 3: if the user enters 3, they wish to draw a triangle (assume right triangle). Ask them for a character to be used to draw the figure:
Enter a character:
Then, ask for a height:
Enter height:
Use this information to draw the triangle. For example, if the user entered the character + and the numbers 4, the rectangle will look like this:
+
++
+++
++++
Option 9: if the user enters 9, they wish to quit the program. Print the goodbye message:
Bye bye!
And terminate the program.
Any other option should produce the message:
Invalid option
Notes/Hints
Every line in output ends with a new line and no white spaces.
You can safely assume that the input will always be valid.
Sample Run:
Please note that this run was produced using a desktop IDE, so the input is visible. In zyBook, the input will not be part of the output of the program and therefore not visible.
Hello, welcome to Figures Drawing!
Select a figure from the menu (or quit):
1 - Square
2 - Rectangle
3 - Triangle
9 - Quit program
1
Enter a character:
&
Enter side length:
4
&&&&
&&&&
&&&&
&&&&
Select a figure from the menu (or quit):
1 - Square
2 - Rectangle
3 - Triangle
9 - Quit program
2
Enter a character:
7
Enter height:
3
Enter width:
8
77777777
77777777
77777777
Select a figure from the menu (or quit):
1 - Square
2 - Rectangle
3 - Triangle
9 - Quit program
7
Invalid option.
Select a figure from the menu (or quit):
1 - Square
2 - Rectangle
3 - Triangle
9 - Quit program
9
Bye bye!